requirements.txt
python-telegram-bot==20.6
python-dotenv
requests
solana
base58
cryptography
mnemonic
bip32
ecdsa
